Add CACHE_ALGORITHM_LRU in BaseCachePolicy
