+++ src/LRUCache.h
@@ -1,23 +1,16 @@
 #pragma once
+#include "CachePolicyAddResult.h"
 
 typedef struct LRUCache_T LRUCache;
 
-// LRUCacheAlloc - create LRUCache structure with hash_table, vector, double_list
-LRUCache* LRUCacheAllocDList(size_t element_size, size_t element_align, size_t value_size, size_t value_align, BaseHashFunc hash_func, BaseCompareFunc compare_func);// +-
+// LRUCacheAlloc - create LRUCache structure with hash_table, double_list, capacity
+LRUCache* lruCacheAlloc(size_t  capacity, size_t element_size, size_t element_align, BaseHashFunc hash_func, BaseCompareFunc compare_func);
+// lruCacheFree - free LRU
+void lruCacheFree(LRUCache* LRU);
 
-// LRUsearchInDoubleList - search element in cache, returns its pointer or NULL
-DoubleListNode* LRUsearchInDoubleList(LRUCache* LRU, void const* key);// +
-
-// LRUCache - LRU algorithm of cache 
-void LRUCache(size_t element_size, size_t element_align, size_t value_size, size_t value_align, BaseHashFunc hash_func, BaseCompareFunc compare_func)
-
-// LRUstep - step of LRU algorithm
-void LRUstep(LRUCache* LRU); // +-
-
-// LRUgetNextData - returns next element from LRU->vector
-void const* LRUgetNextData(LRUCache* LRU); // +-
-
-// lruCacheAdd - adds key into cash and tells wich one remove from cache
-CachePolicyAddResult lruCacheAdd(LRUCache* LRU, void const* key, void const* replace);
-// lruCacheContains - check if the key should added to cache
-bool lruCacheContains(LRUCache* LRU, void const* key);
+// lruCacheAddorReplace - adds key into cash and tells wich one was removed from cache
+CachePolicyAddResult lruCacheAddorReplace(LRUCache* LRU, void const* key, void* replace);
+// lruCacheAdd - adds key into cache
+CachePolicyAddResult lruCacheAdd(LRUCache* LRU, void const* key);
+// lruCacheContains - check if the key should added to cache, returns 1 if yes
+bool lruCacheContains(LRUCache const* LRU, void const* key);
